src/infrastructure/repositories/auth/auth.guard.ts
import {
  CanActivate,
  ExecutionContext,
  forwardRef,
  Inject,
  Injectable,
} from '@nestjs/common';
import { GqlExecutionContext } from '@nestjs/graphql';
import { Observable } from 'rxjs';
import { TokenRepository } from 'src/domain/repositories/token/token.interface';

@Injectable()
export class AuthGuard implements CanActivate {
  constructor(
    @Inject(forwardRef(() => TokenRepository))
    private readonly jwtRepository: TokenRepository,
  ) {}
  canActivate(
    context: ExecutionContext,
  ): boolean | Promise<boolean> | Observable<boolean> {
    const ctx = GqlExecutionContext.create(context);
    const { req } = ctx.getContext();
    const token = req.headers.authorization?.split(' ')[1];

    if (!token) {
      return false;
    }

    return this.jwtRepository.validateToken(token).sub != null;
  }
}
